Rename reporting_units folder to reporting_unit
